Refactor and optimize various modules for improved readability and performance
- Updated DMA controller to use `is_multiple_of` for cycle checks.
- Added `Default` implementations for `Joypad`, `Background`, `Memory`, `Rendering`, `ScreenBuffer`, `Sprites`, `Status`, and `Debugger` structs.
- Simplified condition checks using range patterns for better clarity in `Nes`, `Ppu`, and `EventLoop` modules.
- Enhanced disassembly output formatting in the debugger.
- Improved handling of temporary breakpoints in the event loop.
- Cleaned up shader manager and GL backend for better error handling and readability.
- Fixed minor issues in manual test cartridges for consistency.
- Adjusted assertions in tests for better clarity and consistency.

+++ src/rendering/shader_manager.rs
@@ -157,7 +157,7 @@ impl ShaderManager {
             handle: Some(glow::NativeTexture(
                 std::num::NonZero::new(input_texture).ok_or("Invalid texture ID")?,
             )),
-            format: gl::RGB8 as u32,
+            format: gl::RGB8,
             size: Size::new(256, 240),
         };
 
@@ -166,7 +166,7 @@ impl ShaderManager {
             handle: Some(glow::NativeTexture(
                 std::num::NonZero::new(output_texture).ok_or("Invalid output texture ID")?,
             )),
-            format: gl::RGB8 as u32,
+            format: gl::RGB8,
             size: Size::new(viewport_width, viewport_height),
         };
